+++ samples/golf/src/golf/CollisionMesh.cpp
@@ -64,7 +64,7 @@ void CollisionMesh::updateCollisionMesh(const cro::Mesh::Data& meshData)
     std::int32_t colourOffset = 0;
     for (auto i = 0; i < cro::Mesh::Attribute::Colour; ++i)
     {
-        colourOffset += meshData.attributes[i];
+        colourOffset += static_cast<std::int32_t>(meshData.attributes[i]);
     }
 
 
@@ -77,8 +77,8 @@ void CollisionMesh::updateCollisionMesh(const cro::Mesh::Data& meshData)
     {
         btIndexedMesh groundMesh;
         groundMesh.m_vertexBase = reinterpret_cast<std::uint8_t*>(m_vertexData.data());
-        groundMesh.m_numVertices = meshData.vertexCount;
-        groundMesh.m_vertexStride = meshData.vertexSize;
+        groundMesh.m_numVertices = static_cast<int>(meshData.vertexCount);
+        groundMesh.m_vertexStride = static_cast<int>(meshData.vertexSize);
 
         groundMesh.m_numTriangles = meshData.indexData[i].indexCount / 3;
         groundMesh.m_triangleIndexBase = reinterpret_cast<std::uint8_t*>(m_indexData[i].data());
