+++ samples/golf/src/golf/Weather.cpp
@@ -369,7 +369,7 @@ void GolfState::buildBow()
     glCheck(glBindBuffer(GL_ARRAY_BUFFER, 0));
 
     auto* submesh = &meshData->indexData[0];
-    submesh->indexCount = indices.size();
+    submesh->indexCount = static_cast<std::uint32_t>(indices.size());
     glCheck(glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, submesh->ibo));
     glCheck(glBufferData(GL_ELEMENT_ARRAY_BUFFER, submesh->indexCount * sizeof(std::uint32_t), indices.data(), GL_STATIC_DRAW));
     glCheck(glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, 0));
